requirements.txt
requests
asyncio
aiohttp
loguru
websockets_proxy
fake_useragent==1.5.1
